refactor: move slide shortcuts to composable
